pyproject.toml
[project]
name = "zoomcamp-returns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.5.0",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.10.0",
    "kaggle>=2.0.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
]
